+++ packages/bundles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.16
+
+### Patch Changes
+
+- b8b1d5e4: fix: sourceMap url in prod files but not publish with sourceMap file
+
 ## 0.1.15
 
 ### Patch Changes