+++ packages/miniapp-runtime/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.1.2
+
+### Patch Changes
+
+- b8b1d5e4: fix: sourceMap url in prod files but not publish with sourceMap file
+- Updated dependencies [b8b1d5e4]
+- Updated dependencies [68f19eae]
+  - @ice/runtime@1.2.9
+  - @ice/shared@1.0.2
+
 ## 1.1.1
 
 ### Patch Changes