+++ packages/plugin-miniapp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.1.2
+
+### Patch Changes
+
+- b8b1d5e4: fix: sourceMap url in prod files but not publish with sourceMap file
+- Updated dependencies [b8b1d5e4]
+  - @ice/miniapp-react-dom@1.0.2
+  - @ice/miniapp-runtime@1.1.2
+  - @ice/miniapp-loader@1.1.1
+  - @ice/bundles@0.1.16
+  - @ice/shared@1.0.2
+
 ## 1.1.1
 
 ### Patch Changes